Issue #364 - IE9 whitespace bug fix
In IE9, if you have whitespace and/or carriage returns between table
elements, it causes IE9 to render a blank column and pushes all the
content one column to the right.  It seems the best solution is to
strip it out of the template string when you compile a template.

If you search google for "IE9 renders blank cells" you'll find lots of
references to this problem.