escapeExpression => Handlebars.Utils.escapeExpression
I'm new to handlebars, I just used it in combination with grunt (https://github.com/gruntjs/grunt-contrib-handlebars). The precompiled functions threw "escapeExpression" is not a function - errors.
This change fixed the problem.

But maybe I just used it wrong? I'd appreciate your feedback, if this was indeed a bug or the fault is on my side.