Support async #with contexts
In '{{#with context}}body{{/with}}', when context is a function, the
functions's return value is used as the context in body.  This allows
for computed context, but not when that context is provided
asynchronously, such as when querying a database.

To support asynchronous contexts as well, check if context is a function
that takes arguments.  If it is, call it with the callback to invoke
when done as the first, and #with's hash arguments as the second
argument.  For example, with the following context function:

  person: function(fn) {
    return fn({first: 'Alan', last: 'Johnson'});
  }

The template:

  {{#with person}}{{first}} {{last}}{{/with}}

will render to:

  Alan Johnson