Work around `window` object not being found.
When using Handlebars in a Cloudflare Worker, an environment in which the `window` and `global` objects both don't exist, an error is thrown about `window` being undefined.

In this PR I have slightly changed the code so that if both `window` and `global` do not exist, a Handlebars.noConflict function is generated that
outputs a message saying noConflict is not supported.