Add literal type for `logger.level` string values
By default, `logger.level` is set to "info," type: `string`, even though
types/index.d.ts specifies that `level` should be type: `number`. This
is confusing.

Worse, setting `logger.level` in TypeScript will not allow a string to
be used. A specific number must be provided, even though a string works
because `lookupLevel` converts the given string to the correct number.
So setting `logger.level` must be done in one of two unsatisfactory
ways:

* By using double assertion to lie to TypeScript, i.e., using a string
with `as unknown as number`. This defeats the purpose of types.
* By providing the exact index of the log level desired, which requires
reading Handlebars' logger implementation to see that:
  * `debug` = 0
  * `info` = 1
  * `warn` = 2
  * `error` = 3

  Note that these numbers are not available in the keys that are defined
  in the types (`DEBUG`, `INFO`, `WARN`, and `ERROR`), since they are
  not actually defined in the logger itself.